+++ packages/css/dropdownmenu.css
@@ -1,5 +1,5 @@
 @layer fds.dropdownmenu {
-  .dropdown {
+  .fds-dropdown {
     position: relative;
     padding: var(--fds-spacing-2);
     z-index: 1500;
@@ -10,22 +10,37 @@
     background-color: var(--fds-semantic-background-default);
   }
 
-  .dropdown.small {
+  .fds-dropdown--small {
     min-width: 240px;
     padding: var(--fds-spacing-2);
   }
 
-  .dropdown.medium {
+  .fds-dropdown--medium {
     min-width: 260px;
     padding: var(--fds-spacing-3) var(--fds-spacing-2);
   }
 
-  .dropdown.large {
+  .fds-dropdown--large {
     min-width: 280px;
     padding: var(--fds-spacing-4) var(--fds-spacing-2);
   }
 
-  .dropdown > hr {
+  .fds-dropdown > hr {
     border-color: var(--fds-semantic-border-divider-subtle) !important;
+  }
+
+  .fds-dropdown__item {
+    justify-content: start;
+    padding: 0 var(--fds-spacing-4);
+  }
+
+  .fds-dropdown__section {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+  }
+
+  .fds-dropdown__heading {
+    padding: var(--fds-spacing-2) var(--fds-spacing-4);
   }
 }
