--- packages/react/src/components/DropdownMenu/DropdownMenuGroup/DropdownMenuGroup.module.css
@@ -1,11 +1,0 @@
-@layer fds.dropdownmenu-group {
-  .section {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-  }
-
-  .heading {
-    padding: var(--fds-spacing-2) var(--fds-spacing-4);
-  }
-}
