--- packages/react/src/components/DropdownMenu/DropdownMenuItem/DropdownMenuItem.module.css
@@ -1,6 +1,0 @@
-@layer fds.dropdownmenu-item {
-  .item {
-    justify-content: start;
-    padding: 0 var(--fds-spacing-4);
-  }
-}
