+++ packages/react/src/components/DropdownMenu/index.ts
@@ -1,6 +1,6 @@
 import { DropdownMenu as DropdownRoot } from './DropdownMenu';
-import { DropdownMenuGroup } from './DropdownMenuGroup/DropdownMenuGroup';
-import { DropdownMenuItem } from './DropdownMenuItem/DropdownMenuItem';
+import { DropdownMenuGroup } from './DropdownMenuGroup';
+import { DropdownMenuItem } from './DropdownMenuItem';
 import { DropdownMenuTrigger } from './DropdownMenuTrigger';
 import { DropdownMenuContent } from './DropdownMenuContent';
 
@@ -24,8 +24,8 @@ DropdownMenu.Item.displayName = 'DropdownMenu.Item';
 DropdownMenu.Trigger.displayName = 'DropdownMenu.Trigger';
 
 export type { DropdownMenuProps } from './DropdownMenu';
-export type { DropdownMenuGroupProps } from './DropdownMenuGroup/DropdownMenuGroup';
-export type { DropdownMenuItemProps } from './DropdownMenuItem/DropdownMenuItem';
+export type { DropdownMenuGroupProps } from './DropdownMenuGroup';
+export type { DropdownMenuItemProps } from './DropdownMenuItem';
 export type { DropdownMenuContentProps } from './DropdownMenuContent';
 export {
   DropdownMenu,
